requirements.txt
PyMuPDF
Pillow
pywin32
psutil
pyautogui
jinja2